Refactor fieldmap tab
